refactor: KEEP-1589 extract convergence barrier into standalone module
Move pure convergence barrier logic (buildEdgesByTarget,
signalConvergenceArrival, getReadyDownstreamIds,
propagateConvergenceSkips) into lib/convergence-barrier.ts so the
algorithms can be tested directly rather than through mirrored
re-implementations.

The executor retains a thin executeReadyDownstream wrapper that
calls getReadyDownstreamIds then runs Promise.allSettled.

Tests now import and exercise the real production functions.
